Start Menu so far

+++ src/main/java/com/owary/model/player/Player.java
@@ -63,13 +63,13 @@ public class Player extends GameObject {
         x += velX;
         y += velY;
 
-        x = Game.clamp((int) x, 0, (int)(Game.WIDTH - (width + 5)));
-        y = Game.clamp((int) y, 0, (int)(Game.HEIGHT - (height + 18)));
+        x = Game.clamp( x, 0, (Game.WIDTH - (width + 5)));
+        y = Game.clamp( y, 0, (Game.HEIGHT - (height + 18)));
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle((int) x, (int) y, (int) width, (int) height);
+        return new Rectangle( x,  y,  width,  height);
     }
 
     public int getHealth() {
